Decode percent-encoded characters in search parameter.
So far we only decoded percent-encoded characters in other parameters
than the search parameter. Now we're using our own code for parsing
the query string into a parameter map. This also gives us more control
over whether or not to decode the + sign to a space sign, which we're
not doing anymore with this patch.

Fixes #24384.